+++ packages/preferences/src/browser/util/preference-types.ts
@@ -21,10 +21,10 @@ import {
     CompositeTreeNode as BaseCompositeTreeNode,
     PreferenceInspection,
     CommonCommands,
-    JsonType,
 } from '@theia/core/lib/browser';
 import { Command, MenuPath } from '@theia/core';
 import { JSONValue } from '@theia/core/shared/@phosphor/coreutils';
+import { JsonType } from '@theia/core/lib/common/json-schema';
 
 export namespace Preference {